Video Proxy reference disambiguation
Video Proxy Object - collection of properties and values
vs. 
Video Proxy Element - HTML DOM element
vs. 
Video Proxy Interface - api for accessing an object

Throughout the document there are references to "video proxy" - object and element seem to be used interchangeably. It can create confusion as to whether the Video Player is supposed to provide an actual Proxy Video Element for the SIVIC Ad's Video asset (component) to be loaded and rendered into; OR if the SIVIC spec is declaring that a Proxy Object for the Video Player's own <video> element exists by which the SIVIC Ad's Interactive asset (component) should enter into communication with the Video Player regarding the SIVIC Ad's Video asset status and control – facilitated by the Video Proxy Interface.